+++ components/NavigationDrawer/src/MDCBottomDrawerViewController.h
@@ -27,7 +27,7 @@
  The main content displayed by the drawer.
  Its height is determined by the returned preferred content size.
  */
-@property(nonatomic, nullable) UIViewController *mainContentViewController;
+@property(nonatomic, nullable) UIViewController *contentViewController;
 
 /**
  A header to display above the drawer's main content.
